refactor(client): morphoto表示コンポーネント
