+++ client/src/app/morphoto/_component/morphoto/index.ts
@@ -1,0 +1,1 @@
+export * from "./morphoto";
