+++ client/src/app/morphoto/_component/morphoto/morphoto.css.ts
@@ -1,0 +1,11 @@
+import { style } from "@vanilla-extract/css";
+
+export const morphotoStyle = style({
+  position: "relative",
+  aspectRatio: "1",
+});
+
+export const morphotoImageStyle = style({
+  borderRadius: 10,
+  objectFit: "cover",
+});
