+++ client/src/app/morphoto/_component/morphotos/morphotos.tsx
@@ -1,5 +1,5 @@
 import { readAllMorphotos } from "@/api";
-import Image from "next/image";
+import { Morphoto } from "../morphoto";
 
 import * as styles from "./morphotos.css";
 
@@ -12,16 +12,7 @@ export const Morphotos = async () => {
   return (
     <div className={styles.morphotosStyle}>
       {morphotos.value.data.map((morphoto) => {
-        return (
-          <div key={morphoto.morphoto_id} className={styles.morphotoStyle}>
-            <Image
-              className={styles.morphotoImageStyle}
-              src={"/assets/nijika2.png" /* morphoto.img_url */}
-              fill
-              alt="morphoto"
-            />
-          </div>
-        );
+        return <Morphoto key={morphoto.morphoto_id} morphoto={morphoto} />;
       })}
     </div>
   );
