+++ client/src/types/Morphoto.d.ts
@@ -1,0 +1,5 @@
+export type TMorphoto = {
+  morphoto_id: string;
+  img_url: string;
+  parent_id?: string;
+};
